#291: refactored Copado class into Copado.js
also replaced "classname" with "this"
